pyproject.toml
[project]
name = "housing-boba-index"
version = "0.1.0"
description = "An index correlating housing attractiveness with proximity to top boba shops from Asia."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1",
    "requests>=2.32",
    "python-dotenv>=1.1",
    "geopy>=2.4",
    "redfin>=0.1.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11",
    "pyright>=1.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",    # conflicts with D211 (no-blank-line-before-class)
    "D213",    # conflicts with D212 (multi-line-summary-second-line)
    "COM812",  # conflicts with ruff formatter
    "ISC001",  # conflicts with ruff formatter
    "TC001",   # we use `from __future__ import annotations`; runtime TypedDict imports are fine
    "TC003",   # same rationale as TC001
]

[tool.ruff.lint.per-file-ignores]
"api_clients/demo_data.py" = [
    "E501",    # demo data table lines are intentionally wide for readability
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
